requirements.txt
fastapi==0.114.2
uvicorn==0.30.6
jinja2==3.1.4
python-multipart==0.0.9
pydantic==2.9.2
pydantic-settings==2.4.0
sqlmodel==0.0.22
reportlab==4.2.2
pdfminer.six==20240706
python-docx==1.1.2
httpx==0.27.2
python-dotenv==1.0.1
